Refactored action processing service to include new classes to handle social and business actions
