+++ src/main/resources/database/changes/release-10.51.0/add_sample_unit_type_to_case.sql
@@ -1,0 +1,1 @@
+ALTER TABLE action.case ADD COLUMN sampleunittype varchar(2);
