+++ src/main/resources/database/changes/release-10.51.0/changelog.yml
@@ -9,3 +9,13 @@ databaseChangeLog:
             path: add_sample_unit_id_to_case.sql
             relativeToChangelogFile: true
             splitStatements: false
+
+  - changeSet:
+      id: 10.51.0-2
+      author: Damien Lloyd
+      changes:
+        - sqlFile:
+            comment: Add sample unit type to case table
+            path: add_sample_unit_type_to_case.sql
+            relativeToChangelogFile: true
+            splitStatements: false
